vec ops size i32->u64

+++ icicle_v3/include/icicle/vec_ops.h
@@ -47,13 +47,13 @@ namespace icicle {
   // template APIs
 
   template <typename S>
-  eIcicleError vector_add(const S* vec_a, const S* vec_b, int n, const VecOpsConfig& config, S* output);
+  eIcicleError vector_add(const S* vec_a, const S* vec_b, uint64_t n, const VecOpsConfig& config, S* output);
 
   template <typename S>
-  eIcicleError vector_sub(const S* vec_a, const S* vec_b, int n, const VecOpsConfig& config, S* output);
+  eIcicleError vector_sub(const S* vec_a, const S* vec_b, uint64_t n, const VecOpsConfig& config, S* output);
 
   template <typename S>
-  eIcicleError vector_mul(const S* vec_a, const S* vec_b, int n, const VecOpsConfig& config, S* output);
+  eIcicleError vector_mul(const S* vec_a, const S* vec_b, uint64_t n, const VecOpsConfig& config, S* output);
 
   template <typename S>
   eIcicleError generate_scalars(S* host_scalars, uint64_t size);
@@ -63,13 +63,13 @@ namespace icicle {
 
   // field specific APIs. TODO Yuval move to api headers like icicle V2
   extern "C" eIcicleError CONCAT_EXPAND(FIELD, vector_add)(
-    const scalar_t* vec_a, const scalar_t* vec_b, int n, const VecOpsConfig& config, scalar_t* output);
+    const scalar_t* vec_a, const scalar_t* vec_b, uint64_t n, const VecOpsConfig& config, scalar_t* output);
 
   extern "C" eIcicleError CONCAT_EXPAND(FIELD, vector_sub)(
-    const scalar_t* vec_a, const scalar_t* vec_b, int n, const VecOpsConfig& config, scalar_t* output);
+    const scalar_t* vec_a, const scalar_t* vec_b, uint64_t n, const VecOpsConfig& config, scalar_t* output);
 
   extern "C" eIcicleError CONCAT_EXPAND(FIELD, vector_mul)(
-    const scalar_t* vec_a, const scalar_t* vec_b, int n, const VecOpsConfig& config, scalar_t* output);
+    const scalar_t* vec_a, const scalar_t* vec_b, uint64_t n, const VecOpsConfig& config, scalar_t* output);
 
   extern "C" eIcicleError CONCAT_EXPAND(FIELD, generate_scalars)(scalar_t* host_scalars, uint64_t size);
 
@@ -82,7 +82,7 @@ namespace icicle {
     const Device& device,
     const scalar_t* vec_a,
     const scalar_t* vec_b,
-    int n,
+    uint64_t n,
     const VecOpsConfig& config,
     scalar_t* output)>;
